ui-charts: isolate time related helpers and types
