+++ front/ui/ui-charts/src/common/consts.ts
@@ -1,2 +1,8 @@
 export const PICKING_LAYERS = ['paths', 'overlay'] as const;
 export const LAYERS = ['background', 'graduations', 'paths', 'overlay', 'captions'] as const;
+
+// ========== DURATION CONSTANTS ==========
+
+export const SECOND = 1000;
+export const MINUTE = 60 * SECOND;
+export const HOUR = 60 * MINUTE;
