+++ front/ui/ui-charts/src/common/index.ts
@@ -1,4 +1,8 @@
-export { usePicking, useDraw } from './useCanvas';
+export * from './layers/TimeCaptions';
+export { default as TimeGraduations } from './layers/TimeGraduations';
+
+export { usePicking, useDraw } from './hooks/useCanvas';
+export { getCrispLineCoordinate } from './helpers/time';
 
 export type {
   HoveredItem,
