+++ front/ui/ui-charts/src/spaceTimeChart/index.ts
@@ -10,7 +10,6 @@ export * from './components/WorkScheduleLayer';
 export * from './components/PatternRect';
 export * from './components/Quadrilateral';
 export * from './components/ZoomRect';
-export * from './components/TimeCaptions';
 
 export { DEFAULT_THEME } from './lib/consts';
 export { MouseContext, SpaceTimeChartContext, SpaceTimeChartCanvasContext } from './lib/context';
@@ -26,5 +25,4 @@ export type {
 
 export { isPathOnScreen } from './utils/geometry';
 export { getSpaceAtTime } from './utils/scales';
-export { getCrispLineCoordinate } from './utils/canvas';
 export { computeRectZoomOffsets } from './utils/scales';
