+++ front/ui/ui-charts/src/trackOccupancyDiagram/components/TrackOccupancyStandalone.tsx
@@ -6,9 +6,9 @@ import { TRACK_HEIGHT_CONTAINER } from '../lib/consts';
 import { isOccupancyPickingElement } from './layers/OccupancyZonesLayer';
 import TrackOccupancyCanvas from './TrackOccupancyCanvas';
 import TrackOccupancyManchette from './TrackOccupancyManchette';
+import { HOUR } from '../../common/consts';
 import { Manchette, useManchetteWithSpaceTimeChart } from '../../manchette';
 import { DEFAULT_THEME, SpaceTimeChart } from '../../spaceTimeChart';
-import { HOUR } from '../../spaceTimeChart/lib/consts';
 import type { OccupancyZone, Track } from '../lib/types';
 
 const TrackOccupancyStandalone = ({
